pavilion/chapter-pulse/requirements.txt
slack-sdk>=3.27
anthropic>=0.25
playwright>=1.40
python-dotenv>=1.0
